+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更改日志 | Change Log
 
+## 2024-06-11
+
+- 添加 `PowerShell` 配置文件 `profile.ps1` | Add `PowerShell` config file `profile.ps1`
+
 ## 2024-06-09
 
 - 添加 `prettier` 配置文件 `.prettierrc.toml` | Add `prettier` config file `.prettierrc.toml`